pyproject.toml
[project]
name = "sodoku-backend"
version = "0.1.0"
description = "FastAPI backend for Sudoku game with user authentication"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.23",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    "pydantic[email]>=2.5.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-xdist>=3.3.0",
    "httpx>=0.24.0",
    "faker>=19.0.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
]
